lambda/notifications/nasa-photo-sender/requirements.txt
twilio==9.4.1
python-dotenv==1.0.1
requests==2.32.3